setup: Add classifiers for Python version support
These are used by services like for example caniusepython3.com and requires.io
to check if all dependencies of a project supports Python 3.

I've run the test suite manually on Python 3.3 and 3.4, and added 3.3 to the
Travis CI build. 3.4 should be added to the Travis CI build when Travis starts
supporting 3.4 in the near future.