Use field.id_for_label instead of field.auto_id
This fixes a bug when specifying a custom id for a field and rendering
it with the `|bootstrap` filter.

When using `field.auto_id` in the label's `for` attribute, this is
always set to the default value instead of the custom `id`.

Also fix a typo in the tests: file_fied -> file_field