Added EmailField which has a given domain ending
I added a new Field, which uses the aria functionality described in http://getbootstrap.com/components/#input-groups-basic
This enables to set a fixed domain ending for Email Input. (But only optically. In the form_valid the domain has to be checked again!)
---
This is the second of two edit. We are now complete!